POLIO-88: create messages, split Dashboard
