+++ plugins/polio/js/src/constants/translations/en.json
@@ -1,11 +1,13 @@
 {
-    "iaso.polio.label.campaigns": "Campaigns",
     "iaso.polio.budgetStatus": "Budget status",
+    "iaso.polio.button.label.removeDistrict": "Unselect district",
+    "iaso.polio.button.label.removeRegion": "Unselect region",
     "iaso.polio.calendar": "Calendar",
     "iaso.polio.calendar.obrName": "Name",
     "iaso.polio.calendar.weeks": "week(s)",
     "iaso.polio.form.validator.error.endDateBeforeStartDate": "End date can't be before start date",
     "iaso.polio.form.validator.error.positiveInteger": "Please use a positive integer",
+    "iaso.polio.label.campaigns": "Campaigns",
     "iaso.polio.label.configEmailNotif": "Configure country: {country}",
     "iaso.polio.label.configuration": "Configuration",
     "iaso.polio.raStatus": "RA Status",
