+++ plugins/polio/js/src/constants/translations/fr.json
@@ -1,11 +1,13 @@
 {
-    "iaso.polio.label.campaigns": "Campagnes",
     "iaso.polio.budgetStatus": "Statut budget",
+    "iaso.polio.button.label.removeDistrict": "Déselectionner district",
+    "iaso.polio.button.label.removeRegion": "Déselectionner région",
     "iaso.polio.calendar": "Calendrier",
     "iaso.polio.calendar.obrName": "Nom",
     "iaso.polio.calendar.weeks": "semaine(s)",
     "iaso.polio.form.validator.error.endDateBeforeStartDate": "La date de fin ne peut être avant la date de début",
     "iaso.polio.form.validator.error.positiveInteger": "Veuillez remplir avec un nombre entier positif",
+    "iaso.polio.label.campaigns": "Campagnes",
     "iaso.polio.label.configEmailNotif": "Configurer le pays: {country}",
     "iaso.polio.label.configuration": "Configuration",
     "iaso.polio.raStatus": "Statut RA",
