+++ plugins/polio/js/src/forms/Round2Form.js
@@ -1,0 +1,120 @@
+import React from 'react';
+import { Grid } from '@material-ui/core';
+import { Field } from 'formik';
+import { useSafeIntl } from 'bluesquare-components';
+import { useStyles } from '../styles/theme';
+import MESSAGES from '../constants/messages';
+import { DateInput, TextInput } from '../components/Inputs';
+
+export const Round2Form = () => {
+    const classes = useStyles();
+    const { formatMessage } = useSafeIntl();
+
+    return (
+        <Grid container spacing={2}>
+            <Grid xs={12} md={6} item>
+                <Field
+                    label={formatMessage(MESSAGES.roundTwoStart)}
+                    name="round_two.started_at"
+                    component={DateInput}
+                    fullWidth
+                />
+
+                <Field
+                    label={formatMessage(MESSAGES.roundTwoEnd)}
+                    name="round_two.ended_at"
+                    component={DateInput}
+                    fullWidth
+                />
+                <Field
+                    label={formatMessage(MESSAGES.mopUpStart)}
+                    name="round_two.mop_up_started_at"
+                    component={DateInput}
+                    fullWidth
+                />
+                <Field
+                    label={formatMessage(MESSAGES.mopUpEnd)}
+                    name="round_two.mop_up_ended_at"
+                    component={DateInput}
+                    fullWidth
+                />
+                <Field
+                    label={formatMessage(MESSAGES.imStart)}
+                    name="round_two.im_started_at"
+                    component={DateInput}
+                    fullWidth
+                />
+
+                <Field
+                    label={formatMessage(MESSAGES.imEnd)}
+                    name="round_two.im_ended_at"
+                    component={DateInput}
+                    fullWidth
+                />
+            </Grid>
+            <Grid xs={12} md={6} item>
+                <Field
+                    label={formatMessage(MESSAGES.lqasStart)}
+                    name="round_two.lqas_started_at"
+                    component={DateInput}
+                    fullWidth
+                />
+
+                <Field
+                    label={formatMessage(MESSAGES.lqasEnd)}
+                    name="round_two.lqas_ended_at"
+                    component={DateInput}
+                    fullWidth
+                />
+                <Field
+                    label={formatMessage(MESSAGES.districtsPassingLqas)}
+                    name="round_two.lqas_district_passing"
+                    component={TextInput}
+                    className={classes.input}
+                />
+                <Field
+                    label={formatMessage(MESSAGES.districtsFailingLqas)}
+                    name="round_two.lqas_district_failing"
+                    component={TextInput}
+                    className={classes.input}
+                />
+                <Field
+                    label={formatMessage(MESSAGES.mainReasonForNonVaccination)}
+                    name="round_two.main_awareness_problem"
+                    component={TextInput}
+                    className={classes.input}
+                />
+                <Field
+                    label={formatMessage(
+                        MESSAGES.ratioChildrenMissedInHousehold,
+                    )}
+                    name="round_two.im_percentage_children_missed_in_household"
+                    component={TextInput}
+                    className={classes.input}
+                />
+                <Field
+                    label={formatMessage(
+                        MESSAGES.ratioChildrenMissedOutOfHousehold,
+                    )}
+                    name="round_two.im_percentage_children_missed_out_household"
+                    component={TextInput}
+                    className={classes.input}
+                />
+                <Field
+                    label={formatMessage(
+                        MESSAGES.ratioChildrenMissedInAndOutOfHousehold,
+                    )}
+                    name="round_two.im_percentage_children_missed_in_plus_out_household"
+                    component={TextInput}
+                    className={classes.input}
+                />
+                <Field
+                    label={formatMessage(MESSAGES.awarenessCampaignPlanning)}
+                    name="round_two.awareness_of_campaign_planning"
+                    component={TextInput}
+                    className={classes.input}
+                />
+            </Grid>
+        </Grid>
+    );
+};
